fix go 1.6 build by changing pointer passing
In Go 1.6, we're no longer allowed to pass a pointer to
a Go struct with Go fields to Cairo functions that want a callback.

Work around this by instead making a map keyed by integers and
passing integer keys through to C.  This is a bit cumbersome but
it's the recommended fix for this problem.